backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
mysqlclient==2.2.6
python-dotenv==1.0.1
django-environ==0.11.2
gunicorn==23.0.0